#64 #68 Detailed order page for the seller. Statistics in your personal account. A separate page with detailed statistics

+++ Core/shop/models/orderproducts.py
@@ -7,4 +7,4 @@ class OrderProducts(models.Model):
     product = models.ForeignKey(Product, on_delete=models.CASCADE)
     order = models.ForeignKey(Order, on_delete=models.CASCADE)
     quantity = models.PositiveIntegerField(default=1)
-    price_per_product = models.PositiveIntegerField()
+    price_per_product = models.DecimalField(decimal_places=2, max_digits=10)
